requirements-dev.txt
-r requirements.txt

pytest
pytest-cov
coverage
pyyaml
gitdb
gitpython
twine>=1.11.0
setuptools>=38.6.0
wheel>=0.31.0
